+++ src/cone/app/interfaces.py
@@ -17,7 +17,6 @@ class ISecured(Interface):
 class IApplicationNode(ISecured, INode, IAttributes):
     """Application Node interface.
     """
-    layout = Attribute(u"cone.app.interfaces.ILayout providing object")
     properties = Attribute(u"cone.app.interfaces.IProperties providing object")
     metadata = Attribute(u"cone.app.interfaces.IMetadata implementation")
     nodeinfo = Attribute(u"cone.app.interfaces.INodeInfo providing object")
@@ -61,7 +60,7 @@ class IProperties(IReadMapping):
         """
 
 
-class ILayout(IProperties):
+class ILayoutConfig(IProperties):
     """Layout configuration.
     """
     mainmenu = Attribute(u"Flag whether to display mainmenu")
@@ -73,6 +72,10 @@ class ILayout(IProperties):
     sidebar_left = Attribute(u"Tiles which should be rendered in sidebar")
     sidebar_left_grid_width = Attribute(u"Sidebar grid width")
     content_grid_width = Attribute(u"Content grid width")
+
+
+# B/C, removed as of cone.app 1.1
+ILayout = ILayoutConfig
 
 
 class IMetadata(IProperties):